COMMUNICATION-386 - Waveflow related protocols
Forward port from 8.9

git-svn-id: http://source.eict.vpdc/svn/labo/protocols/trunk@53780 0d108963-8a47-0410-809b-ebd9fb3b3407

+++ upl/com.energyict.mdc.upl.impl/protocolimpl/src/main/java/com/energyict/protocolimpl/coronis/waveflow/core/CommonObisCodeMapper.java
@@ -184,7 +184,7 @@ public class CommonObisCodeMapper {
         try {
             OperatingMode operatingMode = waveFlow.getParameterFactory().readOperatingMode();
             if (obisCode.equals(OBISCODE_REMAINING_BATTERY)) {
-                int level = waveFlow.getParameterFactory().readBatteryLifeDurationCounter().remainingBatteryLife();
+                double level = waveFlow.getParameterFactory().readBatteryLifeDurationCounter().remainingBatteryLife();
                 return new RegisterValue(obisCode, new Quantity(BigDecimal.valueOf(level), Unit.get("")), new Date());
             } else if (obisCode.equals(OBISCODE_APPLICATION_STATUS)) {
                 int status = waveFlow.getParameterFactory().readApplicationStatus();
